+++ src/config/wagmi.ts
@@ -1,11 +1,11 @@
 import { http, createConfig } from 'wagmi'
 import { base } from 'wagmi/chains'
-import { injected } from 'wagmi/connectors'
+import { farcasterMiniApp as miniAppConnector } from '@farcaster/miniapp-wagmi-connector'
 
 export const config = createConfig({
     chains: [base],
     connectors: [
-        injected(),
+        miniAppConnector(),
     ],
     transports: {
         [base.id]: http(),